pyproject.toml
[project]
name = "donna-bot"
version = "0.1.0"
description = "Donna — Harvey's AI concierge, Telegram edition"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0,<22.0",
    "httpx>=0.27.0",
    "pymsalruntime>=0.17.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
